feat:ryusun-product파트 구현

+++ src/main/java/com/example/creditmarket/dto/OrderResponseDTO.java
@@ -10,7 +10,7 @@ public class OrderResponseDTO {
 
     private Long orderId;
 
-    private Long orderStatus;
+    private int orderStatus;
 
     private String fproductCompanyName;
